requirements.txt
pyreadline
python-telegram-bot